# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "description": "Documentation officielle de l'API Shop",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/roles/show/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "description": "Récupérer un rôle par son ID",
                "operationId": "d8bfeea18fb66a5774cf5bfb0ed53b6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rôle récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    }
                }
            }
        },
        "/api/roles/store": {
            "post": {
                "tags": [
                    "Roles"
                ],
                "description": "Créer un rôle",
                "operationId": "c7ec70b3cfebab546f13a2fa82074abe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rôle créé avec succès"
                    },
                    "422": {
                        "description": "Nom déjà existant"
                    },
                    "500": {
                        "description": "Erreur interne serveur"
                    }
                }
            }
        },
        "/api/roles/update/{id}": {
            "put": {
                "tags": [
                    "Roles"
                ],
                "description": "Mettre à jour un rôle",
                "operationId": "c5169b1a63d1fe8ce0225d77df06fe05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Super Admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour réussie"
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    },
                    "422": {
                        "description": "Nom déjà existant"
                    }
                }
            }
        },
        "/api/roles/destroy/{id}": {
            "delete": {
                "tags": [
                    "Roles"
                ],
                "description": "Supprimer un rôle (uniquement si aucun utilisateur lié)",
                "operationId": "a665f671c3d924b32d70dd217d9ab55f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rôle supprimé"
                    },
                    "400": {
                        "description": "Impossible de supprimer le rôle, des utilisateurs y sont liés"
                    },
                    "404": {
                        "description": "Rôle non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}